requirements.txt
# --- Xử lý dữ liệu & Môi trường ---
pandas
python-dotenv
requests
pillow

# --- Não bộ AI (Deep Learning) ---
torch
transformers
scikit-learn

# --- Xử lý tiếng Việt (Bắt buộc cho PhoBERT) ---
pyvi
underthesea

# --- Kết nối Google AI (Data Simulation) ---
google-genai